Upgrade axum to 0.7
Axum 0.6 is not compatible with hyper 1.0, so this upgrades to axum 0.7.

Using cargo-udeps with axum 0.7 requires that we bump the nightly version to at least 2024-04-16.

We also provide a service adaptor to handle that tonic’s Body has diverged from Axum’s body, now that they can’t use a common `hyper` body.