Upgrade axum to 0.7
Axum must be >= 0.7 to support hyper >= 1

Doing this also involves changing the Body type used. Since hyper >= 1 does not provide a generic body type, Axum and tonic both use `BoxBody` to provide a pointer to a Body.

This changes the trait bounds required for methods which accept additional Serivces to be run alongside the primary GRPC service, since those will be routed with Axum, and therefore must accept a BoxBody.

Co-authored-by: Ivan Krivosheev <py.krivosheev@gmail.com>
Co-authored-by: Allan Zhang <allanzhang7@gmail.com>